refactor: removing webpack.common.config.ts
Just moving config into the build config files that still used it.  We can take a pass at making common config later, and probably more granularly.
