--- tools/webpack/webpack.common.config.ts
@@ -1,39 +1,0 @@
-import path from 'path';
-import { Configuration } from 'webpack';
-
-const config: Configuration = {
-  entry: {
-    app: path.resolve(process.cwd(), './src/index'),
-  },
-  output: {
-    path: path.resolve(process.cwd(), './dist'),
-    publicPath: '/',
-  },
-  resolve: {
-    alias: {
-      'site.config': path.resolve(process.cwd(), './site.config'),
-    },
-    extensions: ['.js', '.jsx', '.ts', '.tsx'],
-  },
-  ignoreWarnings: [
-    // Ignore warnings raised by source-map-loader.
-    // some third party packages may ship miss-configured sourcemaps, that interrupts the build
-    // See: https://github.com/facebook/create-react-app/discussions/11278#discussioncomment-1780169
-    /**
-     *
-     * @param {import('webpack').WebpackError} warning
-     * @returns {boolean}
-     */
-    function ignoreSourcemapsloaderWarnings(warning) {
-      return (
-        warning.module
-        // @ts-ignore
-        && warning.module.resource.includes('node_modules')
-        && warning.details
-        && warning.details.includes('source-map-loader')
-      );
-    },
-  ],
-};
-
-export default config;
